backend/requirements.txt
ultralytics>=8.3.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
